refactor: 合并辅助模型设置与视觉识别为单页内部Tab
